make the tokenizer and analyzer regular functions for speed
