+++ index.js
@@ -1,5 +1,4 @@
-// This is a boilerplate file which should not need to be changed.
-module.exports = require('machine').pack({
-  pkg: require('./package.json'),
-  dir: __dirname
-});
+module.exports = {
+  tokenizer: require('./lib/tokenizer'),
+  analyzer: require('./lib/analyzer')
+};
